+++ src/facades.php
@@ -190,9 +190,18 @@ const alert = '\Siler\Monolog\alert';
 const emergency = '\Siler\Monolog\emergency';
 
 
+namespace Siler\Klass;
+const unqualified_name = '\Siler\Klass\unqualified_name';
+
+
 namespace Siler\Str;
 const slugify = '\Siler\Str\slugify';
 const lines = '\Siler\Str\lines';
+const starts_with = '\Siler\Str\starts_with';
+const ends_with = '\Siler\Str\ends_with';
+const contains = '\Siler\Str\contains';
+const snake_case = '\Siler\Str\snake_case';
+const camel_case = '\Siler\Str\camel_case';
 
 
 namespace Siler\Tuple;